@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { EquipeStatistiques } from 'src/app/services/equipe-statistiques';
@@ -6,7 +7,8 @@ import { EquipeStatistiquesService } from 'src/app/services/equipe-statistiques.
 @Component({
   selector: 'app-equipe-statistiques-liste',
   templateUrl: './equipe-statistiques-liste.component.html',
-  styleUrls: ['./equipe-statistiques-liste.component.css']
+  styleUrls: ['./equipe-statistiques-liste.component.css'],
+  imports: [CommonModule]
 })
 export class EquipeStatistiquesListeComponent implements OnInit {
 
